[AVV-4] Create a VERY simplified reedemAda
Problem: We need to somehow demonstrate how an on-chain transaction
worked. One of the proposed ways is to create custom versions of
`redeemAda` and `redeemAdaPaperVend` that don't do any of the
network/database/node stuff and ultimately just prints a transaction,
given the inputs.

Solution: Create `*simple` version of those functions as well as of many
helper versions that depend solely on `MonadThrow`, `MonadCatch`, and
`MonadIO` as the constraints (which are easy to satisfy) and ultimately
just returns a `CTx` without doing any meaningful effects. Note that A
LOT of things were hardcoded to simplify and demonstrate this.